requirements.txt
functions-framework==3.*
requests==2.28.2
google-cloud-storage==2.11.0